Fix csrf token error
We use cookies for CSRF protections with AJAX posts.
But the cookie is only set if a csrf_token template tag is used
somewhere on the page, otherwise we have to decorate the
corresponding view with the `ensure_csrf_cookie` decorator.

Interestingly enough this problem did not occurr earlier.
Apparently, I always opened a page that set the cookie before
doing ajax posts.