Fix copyFilterSet uninitialized selectize error
Changes filter set copy function to first create a new function and
add it to the DOM and only then copy over the values from the old
to the new form.  This fixes the `TypeError: Cannot read property
'addOption'` that was raised when a selectize function of an
uninitialized select element was called.