Fix attrs overriding in dynamic multiselect widget
The custom multiselect widget used to override the render's
`attrs` argument to set a custom css class. This is not a good
idea to start with, and in the current django version this nukes
the widget's css id, which keeps the JS (selectize) plugin from
identifying and initializing the widget.

This commit modifies the custom render to update `attrs` instead
of overriding it.